Refactor file directories & ProjectDetails page
