+++ client/.eslintrc.js
@@ -28,5 +28,6 @@ module.exports = {
     'react/jsx-filename-extension': [1, {extensions: ['.js', '.jsx']}],
     'react/jsx-uses-react': 'off',
     'react/react-in-jsx-scope': 'off',
+    'no-unused-vars': 1,
   },
 };
