+++ client/src/pages/ProjectDetails/index.js
@@ -1,0 +1,3 @@
+import ProjectDetails from './ProjectDetails';
+
+export default ProjectDetails;
