+++ client/src/pages/ServiceDetails/index.js
@@ -1,0 +1,3 @@
+import ServiceDetails from './ServiceDetails';
+
+export default ServiceDetails;
